Viterbi: log probabilities, vocabulary constraints, and bugfixes. - Converted all Viterbi probability interfaces to log probabilities. - Added Viterbi generator wrapper that applies vocabulary constraints. - Fixed bugs: 	- Fixed logits from CANINE being normalised along the wrong dimension (should've been classes, but it was input length!). 	- NumPy datatype for score grid was int, but should've been float. 	- Point probabilities from CANINE had PyTorch gradient tracking and couldn't be converted to NumPy. - Switched CANINE checkpoint.

+++ src/tktkt/models/viterbi/framework.py
@@ -42,7 +42,7 @@ INFTY = float("inf")
 class ViterbiStepScores:
 
     def __init__(self, N: int, K: int, default=0):
-        self.grid = np.full(shape=(N,K), fill_value=default)
+        self.grid = np.full(shape=(N,K), fill_value=default, dtype=np.float32)
 
     def get(self, n: int, k: int):
         return self.grid[n,k]
